Backend/requirements.txt
Django==4.2.9
djoser==2.2.2
django-cors-headers==4.3.1
django-extensions==3.2.3
django-jazzmin==2.6.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
pillow==10.2.0